+++ pandas_ml/skaccessors/test/test_mixture.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python
+import pytest
 
 import sklearn.datasets as datasets
 import sklearn.mixture as mixture
@@ -14,28 +15,21 @@ class TestMixture(tm.TestCase):
         self.assertIs(df.mixture.GaussianMixture, mixture.GaussianMixture)
         self.assertIs(df.mixture.BayesianGaussianMixture,
                       mixture.BayesianGaussianMixture)
-        self.assertIs(df.mixture.DPGMM, mixture.DPGMM)
 
-    def test_Classifications(self):
+    @pytest.mark.parametrize("algo", ['GaussianMixture',
+                                      'BayesianGaussianMixture'])
+    def test_Classifications(self, algo):
         iris = datasets.load_iris()
         df = pdml.ModelFrame(iris)
 
-        models = ['GMM', 'DPGMM', 'VBGMM']
-        for model in models:
-            mod1 = getattr(df.mixture, model)(random_state=self.random_state)
-            mod2 = getattr(mixture, model)(random_state=self.random_state)
+        mod1 = getattr(df.mixture, algo)(random_state=self.random_state)
+        mod2 = getattr(mixture, algo)(random_state=self.random_state)
 
-            df.fit(mod1)
-            mod2.fit(iris.data)
+        df.fit(mod1)
+        mod2.fit(iris.data)
 
-            result = df.predict(mod1)
-            expected = mod2.predict(iris.data)
+        result = df.predict(mod1)
+        expected = mod2.predict(iris.data)
 
-            self.assertIsInstance(result, pdml.ModelSeries)
-            self.assert_numpy_array_almost_equal(result.values, expected)
-
-
-if __name__ == '__main__':
-    import nose
-    nose.runmodule(argv=[__file__, '-vvs', '-x', '--pdb', '--pdb-failure'],
-                   exit=False)
+        self.assertIsInstance(result, pdml.ModelSeries)
+        self.assert_numpy_array_almost_equal(result.values, expected)