+++ pandas_ml/smaccessors/test/test_datasets.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python
+import pytest
 
 import pandas as pd
 import statsmodels.api as sm
@@ -57,7 +58,7 @@ class TestStatsModelsDatasets(tm.TestCase):
     def test_elnino(self):
         data = getattr(sm.datasets.elnino, self.load_method)()
         msg = "Unable to read statsmodels Dataset without exog"
-        with self.assertRaisesRegexp(ValueError, msg):
+        with pytest.raises(ValueError, match=msg):
             pdml.ModelFrame(data)
 
     def test_engel(self):
@@ -84,7 +85,7 @@ class TestStatsModelsDatasets(tm.TestCase):
     def test_macrodata(self):
         data = getattr(sm.datasets.macrodata, self.load_method)()
         msg = "Unable to read statsmodels Dataset without exog"
-        with self.assertRaisesRegexp(ValueError, msg):
+        with pytest.raises(ValueError, match=msg):
             pdml.ModelFrame(data)
 
     def test_modechoice(self):
@@ -97,7 +98,7 @@ class TestStatsModelsDatasets(tm.TestCase):
     def test_nile(self):
         data = getattr(sm.datasets.nile, self.load_method)()
         msg = "Unable to read statsmodels Dataset without exog"
-        with self.assertRaisesRegexp(ValueError, msg):
+        with pytest.raises(ValueError, match=msg):
             pdml.ModelFrame(data)
 
     def test_randhie(self):
@@ -131,7 +132,7 @@ class TestStatsModelsDatasets(tm.TestCase):
     def test_star98(self):
         data = getattr(sm.datasets.star98, self.load_method)()
         msg = 'Data must be 1-dimensional'
-        with self.assertRaisesRegexp(Exception, msg):
+        with pytest.raises(Exception, match=msg):
             pdml.ModelFrame(data)
 
     def test_strikes(self):
@@ -144,7 +145,7 @@ class TestStatsModelsDatasets(tm.TestCase):
     def test_sunspots(self):
         data = getattr(sm.datasets.sunspots, self.load_method)()
         msg = "Unable to read statsmodels Dataset without exog"
-        with self.assertRaisesRegexp(ValueError, msg):
+        with pytest.raises(ValueError, match=msg):
             pdml.ModelFrame(data)
 
     def test_fair(self):
@@ -168,10 +169,12 @@ class TestStatsModelsDatasets(tm.TestCase):
         self.assertEqual(df.target_name, 'murder')
         tm.assert_index_equal(df.data.columns, pd.Index(data.exog_name))
 
+    @pytest.mark.skip('skip until statsmodels GH#4775 is fixed')
     def test_co2(self):
+        # https://github.com/statsmodels/statsmodels/issues/4775
         data = getattr(sm.datasets.co2, self.load_method)()
         msg = "Unable to read statsmodels Dataset without exog"
-        with self.assertRaisesRegexp(ValueError, msg):
+        with pytest.raises(ValueError, match=msg):
             pdml.ModelFrame(data)
 
 
@@ -181,12 +184,9 @@ class TestStatsModelsDatasets_LoadPandas(TestStatsModelsDatasets):
 
     def test_star98(self):
         data = sm.datasets.star98.load_pandas()
-        msg = 'cannot copy sequence with size 2 to array axis with dimension 303'
-        with self.assertRaisesRegexp(Exception, msg):
+        if pdml.compat._PANDAS_ge_023:
+            msg = 'Wrong number of items passed 2, placement implies 303'
+        else:
+            msg = 'cannot copy sequence with size 2 to array axis with dimension 303'
+        with pytest.raises(Exception, match=msg):
             pdml.ModelFrame(data)
-
-
-if __name__ == '__main__':
-    import nose
-    nose.runmodule(argv=[__file__, '-vvs', '-x', '--pdb', '--pdb-failure'],
-                   exit=False)